+++ Base/FUI/renderview.hpp
@@ -27,6 +27,7 @@
 #include <iostream>
 
 #include "fluoglwidget.hpp"
+#include "renderviewLeftTB.hpp"
 
 class RenderView : public QMainWindow
 {
@@ -49,7 +50,14 @@ class RenderView : public QMainWindow
 
     bool getMainWindowStatus();
 
-
+    template<class SpinBoxType, typename V>
+    std::unique_ptr<SpinBoxType> genSpinBox(V floor, V ceiling)
+    {
+      auto newSpinBox = std::make_unique<SpinBoxType>();
+      newSpinBox->setRange(floor,ceiling);
+      newSpinBox->setButtonSymbols(QAbstractSpinBox::NoButtons);
+      return newSpinBox;
+    }
 
   private:
     std::unique_ptr<QToolBar> genToolProp(Qt::Orientation ori);
